+++ src/main/java/com/tutorialsejong/courseregistration/common/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package com.tutorialsejong.courseregistration.common.config;
 
-import com.tutorialsejong.courseregistration.domain.auth.JwtAuthenticationEntryPoint;
-import com.tutorialsejong.courseregistration.domain.auth.JwtAuthenticationFilter;
-import com.tutorialsejong.courseregistration.domain.auth.JwtTokenProvider;
+import com.tutorialsejong.courseregistration.common.security.JwtAuthenticationEntryPoint;
+import com.tutorialsejong.courseregistration.common.security.JwtAuthenticationFilter;
+import com.tutorialsejong.courseregistration.common.security.JwtTokenProvider;
 import com.tutorialsejong.courseregistration.domain.auth.service.CustomUserDetailsService;
 import java.util.Arrays;
 import org.springframework.context.annotation.Bean;
@@ -27,7 +27,8 @@ public class SecurityConfig {
     private final CustomUserDetailsService customUserDetailsService;
     private final JwtAuthenticationEntryPoint unauthorizedHandler;
 
-    public SecurityConfig(JwtTokenProvider tokenProvider,JwtAuthenticationEntryPoint unauthorizedHandler, CustomUserDetailsService customUserDetailsService) {
+    public SecurityConfig(JwtTokenProvider tokenProvider, JwtAuthenticationEntryPoint unauthorizedHandler,
+                          CustomUserDetailsService customUserDetailsService) {
         this.tokenProvider = tokenProvider;
         this.unauthorizedHandler = unauthorizedHandler;
         this.customUserDetailsService = customUserDetailsService;
@@ -44,7 +45,9 @@ public class SecurityConfig {
                 .cors(cors -> cors
                         .configurationSource(request -> {
                             CorsConfiguration config = new CorsConfiguration();
-                            config.setAllowedOrigins(Arrays.asList("https://tutorial-sejong.com", "https://frontend.local.com:3000", "http://localhost:3000"));
+                            config.setAllowedOrigins(
+                                    Arrays.asList("https://tutorial-sejong.com", "https://frontend.local.com:3000",
+                                            "http://localhost:3000"));
                             config.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
                             config.setAllowedHeaders(Arrays.asList("*"));
                             config.setAllowCredentials(true);
@@ -71,7 +74,8 @@ public class SecurityConfig {
     }
 
     @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception {
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration)
+            throws Exception {
         return authenticationConfiguration.getAuthenticationManager();
     }
