+++ src/main/java/com/tutorialsejong/courseregistration/common/security/JwtTokenProvider.java
@@ -1,7 +1,7 @@
-package com.tutorialsejong.courseregistration.domain.auth;
+package com.tutorialsejong.courseregistration.common.security;
 
-import com.tutorialsejong.courseregistration.domain.auth.service.CustomUserDetailsService;
 import com.tutorialsejong.courseregistration.common.exception.JwtAuthenticationException;
+import com.tutorialsejong.courseregistration.domain.auth.service.CustomUserDetailsService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
